refactor: separate file for execute
