+++ api/api.go
@@ -303,61 +303,6 @@ func (vk *VK) RequestUnmarshal(method string, params Params, obj interface{}) er
 	return json.Unmarshal(rawResponse, &obj)
 }
 
-// ExecuteWithArgs a universal method for calling a sequence of other methods
-// while saving and filtering interim results.
-//
-// The Args map variable allows you to retrieve the parameters passed during
-// the request and avoids code formatting.
-//
-// 	return Args.code; // return parameter "code"
-// 	return Args.v; // return parameter "v"
-//
-// https://vk.com/dev/execute
-func (vk *VK) ExecuteWithArgs(code string, params Params, obj interface{}) error {
-	token := vk.AccessToken
-	if vk.IsPoolClient {
-		token = vk.tokenPool.Get()
-	}
-
-	copyParams := make(Params)
-	for key, value := range params {
-		copyParams[key] = FmtValue(value, 0)
-	}
-
-	copyParams["code"] = code
-	copyParams["access_token"] = token
-	copyParams["v"] = vk.Version
-
-	resp, err := vk.Handler("execute", copyParams)
-
-	jsonErr := json.Unmarshal(resp.Response, &obj)
-	if jsonErr != nil {
-		return jsonErr
-	}
-
-	if resp.ExecuteErrors != nil {
-		return &resp.ExecuteErrors
-	}
-
-	return err
-}
-
-// Execute a universal method for calling a sequence of other methods while
-// saving and filtering interim results.
-//
-// https://vk.com/dev/execute
-func (vk *VK) Execute(code string, obj interface{}) error {
-	return vk.ExecuteWithArgs(code, Params{}, obj)
-}
-
-func fmtBool(value bool) string {
-	if value {
-		return "1"
-	}
-
-	return "0"
-}
-
 func fmtReflectValue(value reflect.Value, depth int) string {
 	switch f := value; value.Kind() {
 	case reflect.Invalid:
